serial: use connected notifier only when needed
It's only relevant when the port is open anyway, and marking the port
disconnected in nm_serial_port_close(), which used to be called from
the finalize() function, would trigger the notifier when stuff was
already cleaned up.  So move the nm_serial_port_close() call to
dispose() and remove the connected notifier before we clean the
port up.