gsm: ignore errors from power-on command
Phones especially don't seem to consistently implement this.  For now,
we'll hack it out, but later, we'll want to have a class method for
power-on instead of just a property so that subclasses can decided for
themselves (since they know their hardware better) whether failure
of the power-on command is fatal or not.