longcheer: new plugin for Longcheer (Alcatel etc) devices (bgo #606550)
Anything with vendor ID 0x1c9e really; like Alcatel X020, X030,
X060s, etc.  Longcheer appears to make the actual hardware that all
the devices with vendor ID 0x1c9e use.  You'll see it in .INF files
with "CMLONG" as part of the USB interface definition.

If the ports are not correctly detected, we need to get the driver's
.INF files to determine what the ports should be, and add them to
the udev rules file.