api: revert parts of 2d194a5798fb06f41b018c2d8c2216f45bfc41a4
MM hadn't implemented it yet, but Wader already implemented an earlier
version that didn't use a bitfield but an enum.  Unfortunately the
network mode stuff doesn't allow for distinguishing between the device's
mode preference and the current access technology.  So deprecate the
current network mode stuff in the API in preparation for improved API.