gsm: fix crash in Simple API's GetStatus handling
Can't schedule the info completion until we're sure all the
requests have completed.  They won't necessarily be completed
in the same order they were issued since some of the data the
requests pull from could be cached and thus we don't have to
wait in the queue to hit up the modem.