serial: default to XON/XOFF flow control
This seems to help with Huawei and ZTE devices which often
appeared to stop responding on either primary or secondary ports
at various times.  We had this problem a long time ago, but it was
fixed then by always picking the right serial port via the USB
interface number (Huawei) or udev rules files (ZTE).  Now that we're
using the second serial port more extensively the problem came
up again, so lets try to fix it for real.