gsm: try PDP context deactivation on the second port first
There are some cases where flashing the primary port doesn't work
either due to stupid modem firmware or crappy kernel drivers.  So
if we have a secondary port, try sending the PDP deactivation
command to the secondary port first, and if that fails send it
to the primary port after the primary port gets flashed.  This
increases the chances that the +CGACT request will be successful.

Some modems (Huawei, ZTE) don't like +CGACT on the secondary port,
but when that fails, the code falls back to previous behavior of
flashing and sending CGACT to the primary port.