icera: add complete support for generic Icera devices
Implement connect, disconnect, and IP4 config stuff.

Also fix handling of Icera private data.  After creation we
need to use MM_MODEM_ICERA_GET_INTERFACE(m)->priv to get the
private data instead of just dereferencing the MMModemIcera,
for reasons that I don't know.  If this isn't done, data
gets silently corrupted because writes to the private data
are going into a random location in the object.  This a
side-effect of the slightly hack-ish way that MMModemIcera
is a GInterface with private data.