gsm: send init command twice to make the N900 happy (rh #583691) (lp:765516)
The N900 has some odd serial characteristics in that it appears to
send pieces of the commands back for whatever reason, until you've sent
a few commands down to it.  Almost like it's training on whatever
you send and needs a bit of input to figure out the characteristics.
Whatever.  Just send the init command twice instead of failing when
the N900 barfs the first time.