api: rename 3GPP and CDMA interface paths
Elements in a DBus interface name cannot start with a digit, so
"org.freedesktop.ModemManager1.Modem.3gpp" is an invalid interface name.

Renamed here all relevant interfaces so that they have an additional "Modem"
prefix in the element, so we get now:
 "org.freedesktop.ModemManager1.Modem.Modem3gpp"
 "org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd"
 "org.freedesktop.ModemManager1.Modem.ModemCdma"

Objects generated with gdbus-codegen maintain the previous names.