Revert "huawei: rework probing and detection"
This reverts commit dc89c0a42d826fc3302b3d790d5161945ff7078f.

Older Huawei devices like the E220 really don't like this patch,
but we have a real solution in-store (see the 'ports' branch)
queued up for later.  So until we feel comfortable with that patch,
revert the problem commit.  This will change probing behavior
for newer Huawei devices, but they deal better with secondary
ports anyway, while the E220 is just dumb and so it needs some
special help.