broadband-modem-qmi: implement network scan
The "NAS Network Scan" command replies with several optional TLVs. One of the
TLVs gives the status of the network, using MCC/MNC as index; another one gives
the radio access technology, also using MCC/MNC as index. But, several entries
may have the same MCC/MNC, so there is no clear way to mix the results of both
TLVs in the same element. We'll try to do a best guess, trying not to use the
same RAT result twice, but this is really a best guess.