api: move SupportedBands property to Gsm.Network interface
Gsm.Card is really for the SIM, not for anything to do with bands
or the device or whatever.  Which was a mistake long cleaned up
in MM 0.7+, but not in earlier versions.  Plus, the SupportedBands
and SupportedModes properties were never implemented in any in-tree
drivers.

In the 0.5/0.6 API, SupportedBands belongs on the Network interface
along with the Get/Set methods for bands.  The old property on the
Card interface is deprecated, but not removed, to preserve backwards
compatibility.

The SupportedModes property is also deprecated, but not replaced by
a corresponding property on the Network interface, because it returned
a confusing mix of values (conflating access technology and allowed
mode preference), plus it also was never used by in-tree drivers. If
we really do need a property that advertises the supported allowed
modes of the device (ie 2G_PREFERRED, 2G_ONLY, 3G_ONLY, etc) then
we can add one on the Network interface later.